requirements.txt
Flask>=2.0
Flask-Cors>=3.0
gspread>=5.0
google-auth>=2.0
google-auth-oauthlib>=1.0
oauth2client>=4.1.3
python-dotenv>=1.0
keyring>=23.0
bcrypt>=4.0
requests>=2.0
# Optional (unofficial) library for Google Voice polling/webhooks
google-voice>=0.5.0
